requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
et-xmlfile==1.1.0
idna==3.3
kaleido==0.2.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
pycoingecko==2.2.0
python-dateutil==2.8.1
pytz==2022.1
requests==2.27.1
scipy==1.8.1
six==1.16.0
tenacity==8.0.1
urllib3==1.26.9
eth-brownie>=1.19.0,<2.0.0
